Corriger le calcul du temps pour la detection de terroriste SABETHA
get_time_base() tronquait l'intervalle de polling a l'entier (299 au
lieu de 299.776ms) : cette erreur d'arrondi derive de pres d'un
echantillon complet toutes les ~2.5s, et fait pointer vers la
mauvaise position en fin de combat. C'est ce qui produisait un faux
positif sur Lillith Deros (verifie a la main sur le replay : elle
etait isolee, personne a proximite au moment calcule).

Verification a plusieurs frames autour de bomb_time (au lieu d'un
seul instant) pour lisser le bruit d'echantillonnage a 300ms face a
un porteur en mouvement rapide.

Formule de conversion pixel/pouce revalidee contre le code source
d'Elite Insights (export JSON, export HTML, et le rendu JS des
cercles de portee du viewer officiel : les trois s'accordent sur
game_units = pixel_distance / inchToPixel).

Resultat verifie a la main sur TBQo-20260730-211319_sab via le
cercle de portee natif du site (rayon 280, calibre precisement) :
Schnakao et Wittel Mourndust sont les deux vrais terroristes, pas
Lillith Deros. test_sabetha_terrorist_detection et sab.txt mis a jour
en consequence.

+++ src/models/boss_class.py
@@ -708,9 +708,11 @@ class Boss:
         return 0  
     
     def get_time_base(self):
+        # not rounded to int: truncating ~299.78ms to 299 drifts the index
+        # by almost a full sample every ~2.5s, badly aliasing late-fight lookups
         delta = self.log.pjcontent["players"][0]["combatReplayData"]["end"]-self.log.pjcontent["players"][0]["combatReplayData"]["start"]
         lpos  = len(self.log.pjcontent["players"][0]["combatReplayData"]["positions"])
-        return int(delta/lpos)
+        return delta/lpos
     
     def getBuff(self, buffId: int):
         if not hasattr(self, "_buff_by_id"):
